core_s3: fix every clippy warning in the firmware crate
`cargo clippy --all-targets` on core_s3 (never run before; only `cargo check`
was) reported six warnings, none in the workspace core_s3_* crates:

- lagprobe: `Stamp` was `()` with the feature off, tripping `let_unit_value`
  and `unit_arg` in net.rs. It is now an empty `Copy` struct, so the callers
  stay as they are and no lint is silenced.
- audio.rs / sound.rs: `x % n == 0` const assertions -> `is_multiple_of`.
- platform.rs: drop the `as u64` on a value that is already `u64`.
- audio.rs: `Speaker::rearm` is used by the firmware but not by the
  `sound_test` example that includes the file; allow `dead_code` there.

Clippy is clean for no features, `lagprobe` and `lagprobe-keys`, and
`cargo fmt --check` passes.

+++ rust-doomgeneric/core_s3/src/platform.rs
@@ -57,7 +57,7 @@ fn publish_fps(tenths: u64) {
 }
 
 fn now_us() -> u64 {
-    Instant::now().duration_since_epoch().as_micros() as u64
+    Instant::now().duration_since_epoch().as_micros()
 }
 
 /// Frames counted towards the next published frame rate.
